GifariPhantom/pyproject.toml
[project]
name = "gifariphantom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "pillow>=11.2.1",
    "requests>=2.32.4",
    "ruff>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
